+++ .github/workflows/update_vipb_on_publish_release.yml
@@ -10,7 +10,7 @@ jobs:
     steps:
       - name: Update VIPB Verison
         run: 
-         python ${{runner.workspace}}\grpc-labview\build-it\update_vipb_version.py --library_version ${{github.ref_name}}
+         python ${{runner.workspace}}\grpc-labview\build-it\build.py --library_version ${{github.ref_name}} --build_vipb -1
         continue-on-error: false
     
       - name : Commit VIPB library version bump
